% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult.test.r
\docType{data}
\name{adult.test}
\alias{adult.test}
\title{Validation set of the census income data}
\format{The validation set contains 16,281 observations.}
\description{
Predict whether income exceeds 50K/yr based on census data.
}
\details{
The variables are the same as the traing set (adult.train).
}
\keyword{datasets}

