% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMLE.r
\name{getMLE}
\alias{getMLE}
\title{Calculate the weighted MLE}
\usage{
getMLE(x, y, w)
}
\arguments{
\item{x}{the input covariate matrix}

\item{y}{the input response vector}

\item{w}{the wight vector}
}
\description{
This function calculate the weighted MLE for the input covariate matrix x, response vector y, and weight vector w.
It returns a list with three elements: par, the weighted MLE; msg, the fitting message; iter, the number of itterations used.
}
\examples{
library(OSMAC)
dat <- adult.train
X <- as.matrix(dat[,c(1,3,5,12:13)])
X <- t(t(X) / apply(X, 2, sd))
X <- cbind(1, X)
Y <- as.numeric(dat[,15]) - 1
n <- dim(X)[1]
d <- dim(X)[2]
getMLE(X, Y, 1)
$par
                          Age         Fnlwgt  Education.num   Loss.captial
    -8.6366072      0.6374174      0.0648296      0.8780786      0.2342951
Hours.per.week
     0.5249214

$message
[1] "Successful convergence"

$iter
[1] 5
}
\keyword{getMLE}

