% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMLE.r
\name{getMLE}
\alias{getMLE}
\title{Calculate the weighted MLE}
\usage{
getMLE(x, y, w)
}
\arguments{
\item{x}{the input covariate matrix}

\item{y}{the input response vector}

\item{w}{the wight vector}
}
\description{
This function calculate the weighted MLE for the input covariate matrix x, response vector y, and weight vector w.
It returns a list with three elements: par, the weighted MLE; msg, the fitting message; iter, the number of itterations used.
}
\examples{
library(OSMAC)
dat <- adult.train
X <- as.matrix(dat[,c(1,3,5,12:13)])
X <- t(t(X) / apply(X, 2, sd))
X <- cbind(1, X)
Y <- as.numeric(dat[,15]) - 1
getMLE(X, Y, 1)
}
\keyword{getMLE}

