% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMLE.r
\name{getMLE}
\alias{getMLE}
\title{This function calculate the weighted MLE for the input 
  covariate matrix x, 
  response vector y, 
  and weight vector w.
This function returns a list with three elements: 
  par, the weighted MLE; 
  msg, the fitting message; 
  iter, the number of itterations used.}
\usage{
getMLE(x, y, w)
}
\description{
This function calculate the weighted MLE for the input 
  covariate matrix x, 
  response vector y, 
  and weight vector w.
This function returns a list with three elements: 
  par, the weighted MLE; 
  msg, the fitting message; 
  iter, the number of itterations used.
}

